requirements.txt
yaspin==2.3.0
sh==2.2.2
biopython==1.85
pandas==2.3.2
plotly==6.3.0
nbformat==5.10.4
pydantic==2.11.9
